backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.4.0
gunicorn==21.2.0
python-decouple==3.8
whitenoise==6.6.0
psycopg2-binary==2.9.9